Deployed/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.
gunicorn==20.0.4
dnspython==2.0.0

# C:\Users\Juanig\Desktop\UNS\IAW\proyecto final\Flask-app-1\src\app.py: 1
Flask == 1.1.2

# C:\Users\Juanig\Desktop\UNS\IAW\proyecto final\Flask-app-1\src\app.py: 5
Flask_Compress == 1.5.0

# C:\Users\Juanig\Desktop\UNS\IAW\proyecto final\Flask-app-1\src\app.py: 3
Flask_Cors == 3.0.9

# C:\Users\Juanig\Desktop\UNS\IAW\proyecto final\Flask-app-1\src\app.py: 2
Flask_PyMongo == 2.3.0

# C:\Users\Juanig\Desktop\UNS\IAW\proyecto final\Flask-app-1\src\app.py: 4
Flask_gzip == 0.2

# C:\Users\Juanig\Desktop\UNS\IAW\proyecto final\Flask-app-1\src\helper_functions.py: 1
beautifulsoup4 == 4.9.1

# C:\Users\Juanig\Desktop\UNS\IAW\proyecto final\Flask-app-1\src\scrap_and_cluster_script.py: 7
# C:\Users\Juanig\Desktop\UNS\IAW\proyecto final\Flask-app-1\src\scrap_and_cluster_script_orig.py: 8
networkx == 2.5

# C:\Users\Juanig\Desktop\UNS\IAW\proyecto final\Flask-app-1\src\scrap_and_cluster_script.py: 5
# C:\Users\Juanig\Desktop\UNS\IAW\proyecto final\Flask-app-1\src\scrap_and_cluster_script_orig.py: 6
newspaper3k == 0.2.8

# C:\Users\Juanig\Desktop\UNS\IAW\proyecto final\Flask-app-1\src\helper_functions.py: 5
# C:\Users\Juanig\Desktop\UNS\IAW\proyecto final\Flask-app-1\src\nltkmodules.py: 1
nltk == 3.5

# C:\Users\Juanig\Desktop\UNS\IAW\proyecto final\Flask-app-1\src\app.py: 5
# C:\Users\Juanig\Desktop\UNS\IAW\proyecto final\Flask-app-1\src\helper_functions.py: 10
# C:\Users\Juanig\Desktop\UNS\IAW\proyecto final\Flask-app-1\src\scrap_and_cluster_script.py: 8
# C:\Users\Juanig\Desktop\UNS\IAW\proyecto final\Flask-app-1\src\scrap_and_cluster_script_orig.py: 9
numpy == 1.19.2

# C:\Users\Juanig\Desktop\UNS\IAW\proyecto final\Flask-app-1\src\app.py: 7
# C:\Users\Juanig\Desktop\UNS\IAW\proyecto final\Flask-app-1\src\scrap_and_cluster_script_orig.py: 2
pymongo == 3.11.0

# C:\Users\Juanig\Desktop\UNS\IAW\proyecto final\Flask-app-1\src\scrap_and_cluster_script.py: 3,4
# C:\Users\Juanig\Desktop\UNS\IAW\proyecto final\Flask-app-1\src\scrap_and_cluster_script_orig.py: 4,5
scikit_learn == 0.23.2
